add hadoop configuration utils.
